Update README.md
Added some info on the relationship between this command and Tekton pipelines, as described to me by @garethjevans. I'm not sure if it belongs up where I've put it, or down in the FAQs, but I've put it near the top for now.

I also took the opportunity to suggest some grammatical improvements, which I'm happy to pull out into a separate PR if desired.