fix: commits iteration for semantic strategy
fixes #78

rewriting the way we iterate over commits to ensure that if a commit can't be retrieved or decoded, it won't fail the whole process. instead, we'll just skip this commit. so worst case is that it will bump the patch instead of the major/minor if for some reason a commit can't be read.